Implemented KAR-141, KAR-143, KAR-145. Fixed issues in otehr parts of the code. Testing pending

+++ karedo_persist/src/main/scala/karedo/entity/DbUserKaredos.scala
@@ -18,7 +18,7 @@ case class UserKaredos
 (
   // accountId
   @Key("_id") id: String = UUID.randomUUID().toString
-  , karedos: Int = 0
+  , karedos: Double = 0
   , ts: DateTime = now
 )
 extends Keyable[String]
